+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to `nova-markdown-tui` will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.1] - 17-11-2023
+
+- Fix: Error on console on editor initialization;
+- Fix: decoded entities on markdown URL tags;
+
 ## [1.2.0] - 22-05-2022
 
 - Configurable media upload url;